+++ libvpl/src/mingw_wa.cpp
@@ -1,3 +1,9 @@
+/*############################################################################
+  # Copyright (C) 2020 Intel Corporation
+  #
+  # SPDX-License-Identifier: MIT
+  ############################################################################*/
+
 // temporary workaround for bootstrap link errors on Windows
 extern "C" {
 int clock_gettime(int clk_id, struct timespec *tp) {
@@ -9,16 +15,16 @@ int nanosleep(void *req, void *rem) {
 int BCryptOpenAlgorithmProvider(void *phAlgorithm,
                                 void *pszAlgId,
                                 void *pszImplementation,
-                                unsigned long dwFlags) {
+                                unsigned int dwFlags) {
     return 0;
 }
 int BCryptGenRandom(void *hAlgorithm,
                     unsigned char *pbBuffer,
-                    unsigned long cbBuffer,
-                    unsigned long dwFlags) {
+                    unsigned int cbBuffer,
+                    unsigned int dwFlags) {
     return 0;
 }
-int BCryptCloseAlgorithmProvider(void *hAlgorithm, unsigned long dwFlags) {
+int BCryptCloseAlgorithmProvider(void *hAlgorithm, unsigned int dwFlags) {
     return 0;
 }
 }
